Batch run feature development - Moved the path of Batch Input Nodes
