+++ frontend/src/components/Workspace/FlowChart/FlowChartNode/BatchInputNode/BatchImageFileNode.tsx
