+++ frontend/src/components/Workspace/FlowChart/FlowChartNode/BatchInputNode/BatchMicroscopeFileNode.tsx
